renamve components
