+++ app/components/avo/media_library/item_component.html.erb
