+++ app/components/avo/media_library/item_component.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Avo
+  module MediaLibrary
+    class ItemComponent < Avo::BaseComponent
+      with_collection_parameter :attachment
+
+      prop :attachment, reader: :public
+      prop :display_filename, default: true
+    end
+  end
+end
