+++ app/components/avo/media_library/item_details_component.html.erb
