+++ app/components/avo/media_library/list_component.html.erb
@@ -2,7 +2,7 @@
   <%= c.with_body do %>
     <div class="grid grow-0 min-h-24 gap-x-4 @container/index" style="grid-template-areas: 'stack';">
       <div class="grid grid-cols-1 @sm/index:grid-cols-2 @lg/index:grid-cols-3 @3xl/index:grid-cols-4 @5xl/index:grid-cols-6 gap-4 min-h-0 min-w-0 auto-rows-max p-4" style="grid-area: stack;">
-        <%= render Avo::MediaLibraryItemComponent.with_collection(@attachments) %>
+        <%= render Avo::MediaLibrary::ItemComponent.with_collection(@attachments) %>
       </div>
       <%# TODO: fix the extra margin  %>
       <%= helpers.turbo_frame_tag 'media_library_item_details', class: 'relative empty:hidden bg-white inset-0 w-full h-full block empty:-ml-4 max-h-full', style: 'grid-area: stack;', xsrc: helpers.avo.media_library_item_path(@attachments.first) %>
