--- app/components/avo/media_library_item_component.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Avo
-  class MediaLibraryItemComponent < Avo::BaseComponent
-    with_collection_parameter :attachment
-
-    prop :attachment, reader: :public
-    prop :display_filename, default: true
-  end
-end
