+++ app/views/avo/media_library/items/index.html.erb
@@ -1,1 +1,1 @@
-<%= render Avo::MediaLibraryComponent.new %>
+<%= render Avo::MediaLibrary::ListComponent.new %>
