+++ app/views/avo/media_library/items/show.html.erb
@@ -1,3 +1,3 @@
 <%= turbo_frame_tag 'media_library_item_details' do %>
-  <%= render Avo::MediaLibraryItemDetailsComponent.new(attachment: @attachment) %>
+  <%= render Avo::MediaLibrary::ItemDetailsComponent.new(attachment: @attachment) %>
 <% end %>
